init/requirements.txt
python-keycloak>=7.0
pyyaml
requests
